[NO_ISSUE] roll an abstract class that contains common functionality for AJAX and Web-socket request handlers
